using live backend

+++ lib/models/clientMonth.dart
@@ -23,7 +23,7 @@ class Month {
     Duration _duration = Duration(seconds: value['duration'] ?? 0);
     return Month(
       hourlyRate: (value['mrr'] ?? mrr) / _duration.inHours,
-      mrr: value['mrr'] ?? mrr,
+      mrr: value['mrr'].toDouble() ?? mrr.toDouble(),
       hourlyRateTarget: value['hourlyRateTarget'] ?? 0,
       tags: value['tags'] ?? {},
       duration: _duration,
